api/requirements.txt
fastapi==0.136.0
uvicorn==0.45.0
tensorflow==2.21.0
numpy==2.4.4
pillow==12.2.0
python-multipart==0.0.26
